+++ tests/test_fsic_functions.py
@@ -9,6 +9,7 @@ import unittest
 
 import numpy as np
 
+from fsic.core import VectorContainer
 from fsic.functions import diff, lag, lead
 
 
@@ -47,6 +48,51 @@ class TestFunctions(unittest.TestCase):
                                        np.array([np.nan, 1.0, 1.0, 1.0, 1.0], dtype=float),
                                        equal_nan=True))
         self.assertTrue(np.array_equal(diff(x, 2),
+                                       np.array([np.nan, np.nan, 2.0, 2.0, 2.0], dtype=float),
+                                       equal_nan=True))
+
+
+class TestFunctionsEval(unittest.TestCase):
+
+    def test_eval_lag(self):
+        # Check lag operations in `eval()`
+        container = VectorContainer(range(5))
+        container.add_variable('x', np.arange(5, dtype=float))
+
+        self.assertTrue(np.allclose(container.x, np.array([0.0, 1.0, 2.0, 3.0, 4.0], dtype=float)))
+
+        self.assertTrue(np.array_equal(container.eval('lag(x)'),
+                                       np.array([np.nan, 0.0, 1.0, 2.0, 3.0], dtype=float),
+                                       equal_nan=True))
+        self.assertTrue(np.array_equal(container.eval('lag(x, 2)'),
+                                       np.array([np.nan, np.nan, 0.0, 1.0, 2.0], dtype=float),
+                                       equal_nan=True))
+
+    def test_eval_lead(self):
+        # Check lead operations in `eval()`
+        container = VectorContainer(range(5))
+        container.add_variable('x', np.arange(5, dtype=float))
+
+        self.assertTrue(np.allclose(container.x, np.array([0.0, 1.0, 2.0, 3.0, 4.0], dtype=float)))
+
+        self.assertTrue(np.array_equal(container.eval('lead(x)'),
+                                       np.array([1.0, 2.0, 3.0, 4.0, np.nan], dtype=float),
+                                       equal_nan=True))
+        self.assertTrue(np.array_equal(container.eval('lead(x, 2)'),
+                                       np.array([2.0, 3.0, 4.0, np.nan, np.nan], dtype=float),
+                                       equal_nan=True))
+
+    def test_eval_diff(self):
+        # Check difference operations in `eval()`
+        container = VectorContainer(range(5))
+        container.add_variable('x', np.arange(5, dtype=float))
+
+        self.assertTrue(np.allclose(container.x, np.array([0.0, 1.0, 2.0, 3.0, 4.0], dtype=float)))
+
+        self.assertTrue(np.array_equal(container.eval('diff(x)'),
+                                       np.array([np.nan, 1.0, 1.0, 1.0, 1.0], dtype=float),
+                                       equal_nan=True))
+        self.assertTrue(np.array_equal(container.eval('diff(x, 2)'),
                                        np.array([np.nan, np.nan, 2.0, 2.0, 2.0], dtype=float),
                                        equal_nan=True))