+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 ### Added
 
 - Added new `functions` module for operations on NumPy arrays.
+- Added support for `functions` as standard in `eval()`.
 
 ### Changed
 ### Deprecated